Ammending the last commit.

+++ src/qTESLA-p-I/gauss.c
@@ -159,7 +159,7 @@ static void knuthMergeExchangeG(int32_t a[/*n*/], size_t n)
 }
 
 
-static void kmxGauss(int32_t z[/*CHUNK_SIZE*/], const unsigned char *seed, int nonce) 
+static void kmxGauss(int64_t z[/*CHUNK_SIZE*/], const unsigned char *seed, int nonce) 
 { // Generate CHUNK_SIZE samples from the normal distribution in constant-time
     sdigit_t sampk[(CHUNK_SIZE + CDT_ROWS)*CDT_COLS];
     int32_t sampg[CHUNK_SIZE + CDT_ROWS];
@@ -193,7 +193,7 @@ static void kmxGauss(int32_t z[/*CHUNK_SIZE*/], const unsigned char *seed, int n
     
     // Discard the trailing entries (corresponding to the CDT) and sample the signs
     for (int i = 0; i < CHUNK_SIZE; i++) {
-        z[i] = (sampg[i] << (RADIX32-16)) >> (RADIX32-16);
+        z[i] = (int64_t)((sampg[i] << (RADIX32-16)) >> (RADIX32-16));
     }
 }
